requirements.txt
blis==0.4.1
boto3==1.11.14
botocore==1.14.14
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cymem==2.0.3
docutils==0.15.2
en-core-web-sm==2.2.5
Flask==1.1.1
Flask-Cors==3.0.8
google-search-results==1.7.1
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
murmurhash==1.0.2
newsapi-python==0.2.6
numpy==1.18.1
oauthlib==3.1.0
plac==1.1.3
preshed==3.0.2
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scipy==1.4.1
six==1.13.0
spacy==2.2.3
srsly==1.0.1
thinc==7.3.1
threader==0.1.1
torch==1.4.0
tqdm==4.41.1
tweepy==3.8.0
TwitterAPI==2.5.10
urllib3==1.25.7
wasabi==0.6.0
Werkzeug==0.16.0
zipp==3.0.0
